+++ src/projectgoldstarsx/MoreInformation.java
@@ -11,7 +11,7 @@ public class MoreInformation
     {
         InformationWindow currentSettingsFrame = new InformationWindow("More Information");
         currentSettingsFrame.setLayout(new GridLayout(1, 1));
-        currentSettingsFrame.setSize(900 * ProjectGoldStarsX.multiplier, 500 * ProjectGoldStarsX.multiplier);
+        currentSettingsFrame.setSize(750 * ProjectGoldStarsX.multiplier, 400 * ProjectGoldStarsX.multiplier);
         String output = "Screen Resolution: " + ProjectGoldStarsX.width + "x" + ProjectGoldStarsX.height + "\n"
                 + "Time Project GoldStars X Last Started: " + ProjectGoldStarsX.startDate.toString();
         currentSettingsFrame.addText(ProjectGoldStarsX.bodyText1, output);